Handle non JSON text.
Previous to this PR if we recieved text that was not json, we would
throw an exception and stop logging. This could happen if hapi has a
console.lot in it for instance.

This change catches the JSON.Parse exception, wraps the raw text in a
JSON object, and sets a wasRaw flag that can be filtered against.