Fix: Handle non JSON text (#1)
* Handle non JSON text.

Previous to this PR if we recieved text that was not json, we would
throw an exception and stop logging. This could happen if hapi has a
console.lot in it for instance.

This change catches the JSON.Parse exception, wraps the raw text in a
JSON object, and sets a wasRaw flag that can be filtered against.

* Fix issue found in PR.

use const for arg in the test.
use millisecond time
simplify how we are getting the time.

* Make time be a number not a string.

* Don't add extra nesting to data.

Per Bongo's PR comment.